Updated .gitignore, project packages versions, and Readme.md
The .gitignore file is updated to ignore additional directories such as bin and obj in AuthAndLogin and Repository projects. Also, unnecessary tracking of changes in .idea folder is removed.

Packages versions in AuthAndLogin.csproj and Repository.csproj are updated from a preview version to a stable version 8.0.0.

The structure of AddDbContext statement in the AuthAndLogin/Program.cs section is simplified.

Detailed instructions, code examples, and installation steps are added to the Readme.md to provide more clear and detailed usage of the template for Authentication/Authorization. These changes aim to help new contributors quickly understand how to install and use the project.

Some unnecessary commented code has been removed from the csproj file.